requirements.txt
django
django-reversion
django-taggit
